Implement Titan-grade stack refactoring with registry enums and TypeScript encoding

+++ contracts/interfaces/IUniV2.sol
@@ -1,16 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-/**
- * @title IUniV2
- * @notice Interface for Uniswap V2 style routers (Quickswap, Sushi, etc.)
- */
 interface IUniswapV2Router {
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 }
